use more general types
