+++ wheat-splicer/src/main/scala/x7c1/wheat/splicer/lib/LogMessage.scala
@@ -1,7 +1,5 @@
 package x7c1.wheat.splicer.lib
 
-import sbt.ProcessLogger
-
 import scala.language.implicitConversions
 
 sealed trait LogMessage {
@@ -19,12 +17,12 @@ object LogMessage {
   }
 
   implicit class ReaderLike(message: LogMessage) {
-    def toReader: Reader[HasProcessLogger, Unit] = Reader { context =>
+    def toReader[A <: HasProcessLogger]: Reader[A, Unit] = Reader { context =>
       val logger = context.logger
       message match {
         case _: Error => message.messages foreach (logger.error(_))
         case _: Info => message.messages foreach (logger.info(_))
-        case Multiple(raw) => (raw map (_.toReader)).uniteAll run context
+        case Multiple(raw) => (raw map (_.toReader[A])).uniteAll run context
       }
     }
   }
